requirements.txt
python-gitlab
requests[security]
